Api/WebApi/ExceptionHandler/EntityNullExceptionHandler.cs
using Domain.Exceptions;
using Microsoft.AspNetCore.Diagnostics;
using Microsoft.AspNetCore.Mvc;

namespace WebApi.ExceptionHandler;

public class EntityNullExceptionHandler : IExceptionHandler
{
    public async ValueTask<bool> TryHandleAsync(HttpContext httpContext, Exception exception, CancellationToken cancellationToken)
    {
        if (exception is not EntityNullException) return false;
        httpContext.Response.StatusCode = StatusCodes.Status404NotFound;
        string details = ((EntityNullException)exception).Message;

        await httpContext.Response.WriteAsJsonAsync(new ProblemDetails()
        {
            Status = httpContext.Response.StatusCode,
            Detail = details,
        });
        return true;
    }
}